+++ src/Basis/wls.jl
@@ -7,12 +7,13 @@ This results in the same kernel used in moving least squares MPM[^MLSMPM].
 
 [^MLSMPM]: [Hu, Y., Fang, Y., Ge, Z., Qu, Z., Zhu, Y., Pradhana, A. and Jiang, C., 2018. A moving least squares material point method with displacement discontinuity and two-way rigid body coupling. ACM Transactions on Graphics (TOG), 37(4), pp.1-14.](https://doi.org/10.1145/3197517.3201293)
 """
-struct WLS{K <: Kernel, P <: Polynomial} <: Basis
+struct WLS{K <: Kernel, P <: CorrectionPolynomial} <: Basis
     kernel::K
     poly::P
 end
 
 WLS(k::Kernel) = WLS(k, Polynomial(Linear()))
+WLS(::Kernel, poly::Polynomial) = throw(ArgumentError(unsupported_correction_polynomial("WLS", poly)))
 
 support_width(wls::WLS) = support_width(wls.kernel)
 @inline supportnodes(wls::WLS, pt, mesh::CartesianMesh) = supportnodes(wls.kernel, pt, mesh)